added vault assets list to vautl details
