--- packages/web/src/pages/Honeypots/VaultDetailsPage/Sections/VaultSubmissionsSection/VaultSubmissionsSection.tsx
@@ -1,9 +1,0 @@
-import { useContext } from "react";
-import { VaultDetailsContext } from "../../store";
-import { StyledSubmissionsSection } from "./styles";
-
-export const VaultSubmissionsSection = () => {
-  const { vault } = useContext(VaultDetailsContext);
-
-  return <StyledSubmissionsSection>VaultSubmissionsSection</StyledSubmissionsSection>;
-};
