--- packages/web/src/pages/Honeypots/VaultDetailsPage/Sections/VaultSubmissionsSection/styles.ts
@@ -1,3 +1,0 @@
-import styled from "styled-components";
-
-export const StyledSubmissionsSection = styled.div``;
