+++ packages/web/src/pages/Honeypots/VaultDetailsPage/Sections/index.ts
@@ -1,4 +1,3 @@
 export { VaultRewardsSection } from "./VaultRewardsSection/VaultRewardsSection";
 export { VaultScopeSection } from "./VaultScopeSection/VaultScopeSection";
 export { VaultDepositsSection } from "./VaultDepositsSection/VaultDepositsSection";
-export { VaultSubmissionsSection } from "./VaultSubmissionsSection/VaultSubmissionsSection";
